[muzi] refactorting by moving code into src :fist:
